multimedia/tvheadend: Add net/uriparser. Re-add ffmpeg transcoder.
This is not merged upstream yet. It fixes transcoding in tvheadend to work for the current ffmpeg version in FreeBSD ports tree (ffmpeg-2.2.4).

Option is now called "FFMPEG" because libav's libs isn't detected by tvheadend, since libav port installs into a subdir to not conflict with ffmpeg. No reason not to use ffmpeg instead. It is same number of dependancies.

A small group of upstream contributors are now trying to fix the remaining libavcodec breakage, so tvheadend can also work with the newer / newest ffmpeg-2.3.2. That is still not quite done yet.